refactor: mediator, ol, and validate
